feat: add prisma db push to prebuild scrip
This commit updates the prebuild script in the package.json file. The script now runs both 'prisma generate' and 'prisma db push' before the application is built. This ensures that the Prisma Client is generated and the Prisma schema is synced with the database, which is particularly important when deploying on platforms like Vercel that cache dependencies. This change helps to prevent the PrismaClientKnownRequestError.